+++ Automate/Framework/GenericModConfigMenuIntegrationForAutomate.cs
@@ -5,6 +5,9 @@ using Pathoschild.Stardew.Automate.Framework.Models;
 using Pathoschild.Stardew.Common;
 using Pathoschild.Stardew.Common.Integrations.GenericModConfigMenu;
 using StardewModdingAPI;
+using StardewValley;
+using StardewValley.GameData;
+using StardewValley.TerrainFeatures;
 
 namespace Pathoschild.Stardew.Automate.Framework
 {
@@ -96,15 +99,16 @@ namespace Pathoschild.Stardew.Automate.Framework
 
             // connectors
             menu.AddSectionTitle(I18n.Config_Title_Connectors);
-            foreach (DataModelFloor entry in this.Data.FloorNames.Values)
+            foreach (FloorPathData entry in Flooring.GetFloorPathLookup().Values.OrderBy(p => GameI18n.GetObjectName(p.ItemID), StringComparer.OrdinalIgnoreCase)) // sort by English display name; it's not ideal, but we can't re-sort when the language is loaded
             {
-                string itemId = entry.ItemId;
+                string itemId = entry.ItemID;
+                string internalName = Utility.GetItemNameForItemID(entry.ItemID);
 
                 menu.AddCheckbox(
                     name: () => GameI18n.GetObjectName(itemId),
                     tooltip: () => I18n.Config_Connector_Desc(itemName: GameI18n.GetObjectName(itemId)),
-                    get: config => this.HasConnector(config, entry.Name),
-                    set: (config, value) => this.SetConnector(config, entry.Name, value)
+                    get: config => this.HasConnector(config, internalName),
+                    set: (config, value) => this.SetConnector(config, internalName, value)
                 );
             }
             menu.AddTextbox(
@@ -138,13 +142,15 @@ namespace Pathoschild.Stardew.Automate.Framework
         /****
         ** Connectors
         ****/
-        /// <summary>Get whether the given item name isn't one of the connectors listed in <see cref="DataModel.FloorNames"/>.</summary>
+        /// <summary>Get whether the given item name isn't one of the connectors listed in <see cref="Flooring.GetFloorPathLookup"/>.</summary>
         /// <param name="name">The item name.</param>
         private bool IsCustomConnector(string name)
         {
-            foreach (DataModelFloor floor in this.Data.FloorNames.Values)
+            foreach (FloorPathData floor in Flooring.GetFloorPathLookup().Values)
             {
-                if (string.Equals(floor.Name, name, StringComparison.OrdinalIgnoreCase))
+                string internalName = Utility.GetItemNameForItemID(floor.ItemID);
+
+                if (string.Equals(internalName, name, StringComparison.OrdinalIgnoreCase))
                     return false;
             }
 
@@ -212,7 +218,6 @@ namespace Pathoschild.Stardew.Automate.Framework
         }
 
         /// <summary>Get the default override for a mod, if any.</summary>
-        /// <param name="config">The mod configuration.</param>
         /// <param name="name">The machine name.</param>
         public ModConfigMachine GetDefaultOverride(string name)
         {
@@ -240,6 +245,7 @@ namespace Pathoschild.Stardew.Automate.Framework
         /// <summary>Get the custom override for a mod, if any.</summary>
         /// <param name="config">The mod configuration.</param>
         /// <param name="name">The machine name.</param>
+        /// <param name="enabled">Whether the custom override should be enabled; else disabled.</param>
         public void SetCustomOverride(ModConfig config, string name, bool enabled)
         {
             ModConfigMachine options = this.GetCustomOverride(config, name) ?? new ModConfigMachine { Enabled = enabled };